+++ server/api/blog/model/request/role.go
@@ -1,16 +1,16 @@
 package request
 
 type UpdateUserRolesReq struct {
-	UserId  int   `json:"user_id"`
-	RoleIds []int `json:"role_ids"`
+	UserId  int64   `json:"user_id"`
+	RoleIds []int64 `json:"role_ids"`
 }
 
 type UpdateRoleMenusReq struct {
-	RoleId  int   `json:"role_id"`
-	MenuIds []int `json:"menu_ids"`
+	RoleId  int64   `json:"role_id"`
+	MenuIds []int64 `json:"menu_ids"`
 }
 
 type UpdateRoleApisReq struct {
-	RoleId      int   `json:"role_id"`
-	ResourceIds []int `json:"resource_ids"`
+	RoleId      int64   `json:"role_id"`
+	ResourceIds []int64 `json:"resource_ids"`
 }
